Redesign Login Activity
